Refactor: #40 - 영화세부 VC 파일 분리
